Move meilisearch http entities package
