+++ src/main/java/de/skillkiller/documentdbackend/controller/DocumentController.java
@@ -5,9 +5,9 @@ import de.skillkiller.documentdbackend.entity.Document;
 import de.skillkiller.documentdbackend.entity.User;
 import de.skillkiller.documentdbackend.entity.UserDetailsHolder;
 import de.skillkiller.documentdbackend.entity.http.frontend.response.UpdateDocumentResponse;
+import de.skillkiller.documentdbackend.entity.http.meilisearch.response.SearchResponse;
 import de.skillkiller.documentdbackend.search.DocumentSearch;
 import de.skillkiller.documentdbackend.search.UserSearch;
-import de.skillkiller.documentdbackend.search.response.SearchResponse;
 import de.skillkiller.documentdbackend.service.AccessTokenService;
 import de.skillkiller.documentdbackend.task.PDFOCR;
 import de.skillkiller.documentdbackend.util.FileUtil;
