+++ src/main/java/de/skillkiller/documentdbackend/entity/http/meilisearch/request/CreateIndexRequest.java
@@ -1,4 +1,4 @@
-package de.skillkiller.documentdbackend.search.request;
+package de.skillkiller.documentdbackend.entity.http.meilisearch.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
