+++ src/main/java/de/skillkiller/documentdbackend/entity/http/meilisearch/response/SearchResponse.java
@@ -1,4 +1,4 @@
-package de.skillkiller.documentdbackend.search.response;
+package de.skillkiller.documentdbackend.entity.http.meilisearch.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
