+++ src/main/java/de/skillkiller/documentdbackend/search/DocumentSearch.java
@@ -2,7 +2,7 @@ package de.skillkiller.documentdbackend.search;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import de.skillkiller.documentdbackend.entity.Document;
-import de.skillkiller.documentdbackend.search.response.SearchResponse;
+import de.skillkiller.documentdbackend.entity.http.meilisearch.response.SearchResponse;
 import kong.unirest.HttpResponse;
 import kong.unirest.JsonNode;
 import kong.unirest.Unirest;
